Avoid decoding results from pyvo in IVOA queries
Recent pyvo (Astropy, etc) returns strings, rather than bytes. Trying to
decode them will raise an exception.

Previously, the exceptions were (usually) masked, as we had them wrapped
inside a try-except, so it wasn't obvious that this was broken. But broken it
was, and it would fail in pathological circumstances.