Use a simple query serializer
The dyanmic query serializer assumes that all results have the same fields,
but this isn't always the case with Zenodo. However, we also don't need a
complex variety of fields, so the dynamism isn't really a plus: we can just
directly specify the data we need. Simpler and more robust!